chore: extract Script commands into separate processor
