+++ src/bidiMapper/domains/script/PreloadScriptStorage.ts
@@ -14,8 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import type {CdpTarget} from '../context/cdpTarget.js';
+
 import type {BidiPreloadScript} from './bidiPreloadScript.js';
-import type {CdpTarget} from './cdpTarget.js';
 
 /** BidiPreloadScripts can be filtered by BiDi ID or target ID. */
 export type BidiPreloadScriptFilter = Partial<
